Move command related classes to core.command package.
