+++ src/core/command/PartyMemberRemoveCommand.java
@@ -1,11 +1,12 @@
-package core.utils;
+package core.command;
 
+import core.command.ICommand;
 import core.session.SessionInformation;
 
 /**
  * Created by David on 23/10/2016.
  */
-public class PartyMemberRemoveCommand implements ICommand{
+public class PartyMemberRemoveCommand implements ICommand {
     private final String name;
     private int memberId;
