+++ src/core/command/PlayerLogOutCommand.java
@@ -1,11 +1,12 @@
-package core.utils;
+package core.command;
 
+import core.command.ICommand;
 import core.session.SessionInformation;
 
 /**
  * Created by David on 23/10/2016.
  */
-public class PlayerLogOutCommand implements ICommand{
+public class PlayerLogOutCommand implements ICommand {
     private final String name;
 
     /**
