+++ src/userinterface/MainMenu.java
@@ -8,14 +8,12 @@
 **/
 package userinterface;
 
-import core.utils.SessionController;
+import core.command.SessionController;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.FlowLayout;
 import java.awt.GridLayout;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import javax.swing.JButton;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
